fix: restore missing media helpers and inject pyflakes static check into tests
